+++ src/App.js
@@ -1,32 +1,11 @@
 import './App.css';
 import { getFacturas } from './API/requests';
-import Modal from '@mui/material/Modal';
-import Box from '@mui/material/Box';
-import Paper from '@mui/material/Paper';
-import Typography from '@mui/material/Typography';
 import { Container, Button } from "@mui/material";
 import { useState, useEffect } from 'react';
 import FacturasComponent from "./Components/Facturas";
 import CreditNoteComponent from "./Components/CreditNotes";
+import ModalComponent from './Components/Modal';
 
-const modalStyle = {
-  position: 'absolute',
-  top: '50%',
-  left: '50%',
-  transform: 'translate(-50%, -50%)',
-  width: 300, // Ancho personalizable
-  bgcolor: 'background.paper',
-  boxShadow: 24,
-  p: 4,
-};
-
-function TicketHeader() {
-  return (
-    <div style={{ textAlign: 'center' }}>
-      ✅
-    </div>
-  );
-}
 
 function App() {
   const [facturas, setFacturas] = useState([]);
@@ -86,30 +65,13 @@ function App() {
             <Button variant="contained" color="primary" onClick={handleOpen}>
               Asignar
             </Button>
-
-            <Modal open={modalOpen} onClose={handleClose}>
-              <Box sx={modalStyle}>
-                <Paper>
-                  <TicketHeader />
-                  <Typography id="modal-modal-title" variant="h6" component="h2">
-                    Nota de crédito asignada correctamente
-                  </Typography>
-
-                  <Button
-                    sx={{ mt: 3 }}
-                    variant="contained"
-                    color="primary"
-                    onClick={handleClose}
-                  >
-                    Seguir asignando
-                  </Button>
-                </Paper>
-              </Box>
-            </Modal>
+            <ModalComponent
+              modalOpen={modalOpen}
+              handleClose={handleClose} />
           </div>
         ) : null}
       </div>
-    </Container>
+    </Container >
   );
 }
